sse.gg/[username] unfurls as a blink
