+++ src/components/common/follow-blink-button.tsx
@@ -28,9 +28,7 @@ export function FollowBlinkButton({
     if (typeof window === 'undefined') return ''
     const { protocol, host } = window.location
     // Updated to use dynamic route structure with username as path parameter
-    return `${protocol}//${host}/api/actions/follow/${encodeURIComponent(
-      username
-    )}`
+    return `${protocol}//${host}/${encodeURIComponent(username)}`
   }
 
   const handleCopyBlinkUrl = async () => {
